Refactor oauth authorized apps endpoint
