--- components/org.wso2.carbon.identity.api.user.authorized.apps/org.wso2.carbon.identity.api.user.authorized.apps.v1/src/gen/java/org/wso2/carbon/identity/api/user/authorized/apps/v1/factories/MeApiServiceFactory.java
@@ -1,13 +1,0 @@
-package org.wso2.carbon.identity.api.user.authorized.apps.v1.factories;
-
-import org.wso2.carbon.identity.api.user.authorized.apps.v1.MeApiService;
-
-public class MeApiServiceFactory {
-
-   private final static MeApiService service = new MeApiServiceImpl();
-
-   public static MeApiService getMeApi()
-   {
-      return service;
-   }
-}
