--- components/org.wso2.carbon.identity.api.user.authorized.apps/org.wso2.carbon.identity.api.user.authorized.apps.v1/src/gen/java/org/wso2/carbon/identity/api/user/authorized/apps/v1/factories/UserIdApiServiceFactory.java
@@ -1,14 +1,0 @@
-package org.wso2.carbon.identity.api.user.authorized.apps.v1.factories;
-
-import org.wso2.carbon.identity.api.user.authorized.apps.v1.UserIdApiService;
-import org.wso2.carbon.identity.api.user.authorized.apps.v1.impl.UserIdApiServiceImpl;
-
-public class UserIdApiServiceFactory {
-
-   private final static UserIdApiService service = new UserIdApiServiceImpl();
-
-   public static UserIdApiService getUserIdApi()
-   {
-      return service;
-   }
-}
