+++ components/org.wso2.carbon.identity.api.user.authorized.apps/pom.xml
@@ -18,20 +18,21 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
 
     <parent>
+        <artifactId>identity-api-user</artifactId>
         <groupId>org.wso2.carbon.identity.user.api</groupId>
-        <artifactId>identity-user-api</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../..</relativePath>
+        <version>1.0.9-SNAPSHOT</version>
+        <relativePath>../../pom.xml</relativePath>
     </parent>
+    <modelVersion>4.0.0</modelVersion>
 
+    <groupId>org.wso2.carbon.identity.user.api</groupId>
     <artifactId>org.wso2.carbon.identity.api.user.authorized.apps</artifactId>
     <packaging>pom</packaging>
 
     <modules>
-        <module>org.wso2.carbon.identity.api.user.authorized.apps.v1</module>
+        <module>org.wso2.carbon.identity.rest.api.user.authorized.apps.v1</module>
     </modules>
     
 </project>
